+++ CHANGELOG.md
@@ -1,3 +1,30 @@
+## [1.14.1](https://github.com/softwaregroup-bg/ut-front-devextreme/compare/v1.14.0...v1.14.1) (2021-05-24)
+
+
+
+# [1.14.0](https://github.com/softwaregroup-bg/ut-front-devextreme/compare/v1.13.0...v1.14.0) (2021-05-22)
+
+
+### Bug Fixes
+
+* wrap prime [UTCORE-121] ([1c44072](https://github.com/softwaregroup-bg/ut-front-devextreme/commit/1c44072e11d9c3ffaf6f89eba7b795dd20b7892c))
+
+
+### Features
+
+* external submit [UTCORE-121] ([f626b5a](https://github.com/softwaregroup-bg/ut-front-devextreme/commit/f626b5a94a393c2267549ee8b7fbf440e6143093))
+
+
+
+# [1.13.0](https://github.com/softwaregroup-bg/ut-front-devextreme/compare/v1.12.1...v1.13.0) (2021-05-21)
+
+
+### Features
+
+* add Editor and ThumbIndex [UTCORE-121] ([382315e](https://github.com/softwaregroup-bg/ut-front-devextreme/commit/382315ed21e99013383404b31e0c20170e5578f3))
+
+
+
 ## [1.12.1](https://github.com/softwaregroup-bg/ut-front-devextreme/compare/v1.12.0...v1.12.1) (2021-05-17)